+++ app-backend/batch/src/main/scala/groundwork/NotifyIntercom.scala
@@ -1,6 +1,8 @@
 package com.rasterfoundry.batch.groundwork
 
 import com.rasterfoundry.batch.Job
+import com.rasterfoundry.database.util.RFTransactor
+import com.rasterfoundry.notification.intercom.IntercomConversation
 import com.rasterfoundry.notification.intercom.Model._
 import com.rasterfoundry.notification.intercom._
 
@@ -17,25 +19,33 @@ object NotifyIntercomProgram extends Job {
     backend <- backendRef
   } yield backend
 
-  def runJob(args: List[String]): IO[Unit] = args match {
-    case externalId +: msg +: Nil =>
-      for {
-        backend <- getBackend
-        _ <- {
-          val notifier = new LiveIntercomNotifier[IO](backend)
-          notifier.notifyUser(
-            Config.intercomToken,
-            Config.intercomAdminId,
-            ExternalId(externalId),
-            Message(msg)
+  def runJob(args: List[String]): IO[Unit] = {
+    val xa =
+      RFTransactor.nonHikariTransactor(RFTransactor.TransactorConfig())
+
+    val dbIO = new DbIO(xa);
+
+    args match {
+      case externalId +: msg +: Nil =>
+        for {
+          backend <- getBackend
+          _ <- {
+            IntercomConversation.notifyIO(
+              externalId,
+              Message(msg),
+              dbIO.groundworkConfig,
+              new LiveIntercomNotifier[IO](backend),
+              dbIO.getConversation,
+              dbIO.insertConversation
+            )
+          }
+        } yield ()
+      case _ =>
+        IO.raiseError(
+          new Exception(
+            s"Arguments should match pattern `USER_ID MESSAGE`. Got $args"
           )
-        }
-      } yield ()
-    case _ =>
-      IO.raiseError(
-        new Exception(
-          s"Arguments should match pattern `USER_ID MESSAGE`. Got $args"
         )
-      )
+    }
   }
 }