+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for Intercom conversations [#5571](https://github.com/raster-foundry/raster-foundry/pull/5571)
 
 ## [1.62.1] - 2021-04-19
 ### Fixed